+++ core/common/src/main/kotlin/com/twobits/common/di/DispatchersModule.kt
@@ -1,0 +1,22 @@
+package com.twobits.common.di
+
+import com.twobits.common.AppDispatcher
+import com.twobits.common.AppDispatchers
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.Dispatchers
+
+@Module
+@InstallIn(SingletonComponent::class)
+object DispatchersModule {
+    @Provides
+    @AppDispatcher(AppDispatchers.IO)
+    fun providesIODispatcher(): CoroutineDispatcher = Dispatchers.IO
+
+    @Provides
+    @AppDispatcher(AppDispatchers.Default)
+    fun providesDefaultDispatcher(): CoroutineDispatcher = Dispatchers.Default
+}